+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.29:
 
 * Add guard against adding file that does not exist to GlassFishContainer's classpath.
+* Add an extra constructor to GlassFishContainer for usability purposes.
 
 ## 0.28: